feat get product by id route

+++ src/services/Products/delete.service.ts
@@ -5,7 +5,11 @@ import { AppError } from '../../errors/AppError';
 export const deleteProductService = async (id: number) => {
     const productRepository = AppDataSource.getRepository(Product)
 
-    const product = await productRepository.findOneByOrFail({ id: id })
+    const product = await productRepository.findOneBy({ id: id })
+
+    if (!product) {
+        throw new AppError(`Product with ID ${id} not found`, 404)
+    }
     
     await productRepository.delete({ id: id })
